+++ src/main/java/com/example/community_service/category/infrastructure/MainCategoryRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MainCategoryRepository extends JpaRepository<MainCategory, Integer> {
+public interface MainCategoryRepository extends JpaRepository<MainCategory, Long> {
 }
